requirement.txt
docker>=5.0.3
PyYAML<6,>=3.10